+++ husky_components/datagrid/main.js
@@ -1007,7 +1007,7 @@
                 this.currentUrl = this.getUrl(params);
 
                 if (!params.data) params.data = {};
-                params.data['expand'] = this.getSelectedItemIds().join(',');
+                params.data['expandIds'] = this.getSelectedItemIds().join(',');
 
                 this.sandbox.dom.addClass(this.$find('.selected-elements'), 'invisible');
                 return this.sandbox.util.load(this.currentUrl, params.data)
@@ -2262,7 +2262,7 @@
              * @param recordId {Number|String}
              */
             loadChildren: function(recordId) {
-                var def = this.sandbox.data.deferred();
+                var deferred = this.sandbox.data.deferred();
 
                 if (!!this.data.links.children) {
                     var template = this.sandbox.uritemplate.parse(this.data.links.children.href),
@@ -2271,16 +2271,17 @@
                     this.sandbox.util.load(this.getUrl({url: url}))
                         .then(function(response) {
                             this.addRecordsHandler(response._embedded[this.options.resultKey]);
-                            def.resolve();
+                            deferred.resolve();
                         }.bind(this))
                         .fail(function(status, error) {
                             this.sandbox.logger.error(status, error);
+                            deferred.fail();
                         }.bind(this));
                 } else {
-                    def.resolve();
+                    deferred.resolve();
                 }
 
-                return def;
+                return deferred;
             },
 
             /**